requirements.txt
absl-py==1.4.0
appdirs==1.4.4
arandr==0.1.10
astroid==2.5.1
asttokens==2.0.4
astunparse==1.6.3
audioread==3.0.0
automationhat==0.2.0
beautifulsoup4==4.9.3
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
cachetools==5.3.1
Cap1xxx==0.1.3
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coloredlogs==15.0.1
colorzero==1.1
contourpy==1.1.0
cryptography==3.3.2
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.21
dbus-python==1.2.16
decorator==5.1.1
distro==1.5.0
docutils==0.16
drumhat==0.1.0
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==1.1.2
flatbuffers==20181003210633
fonttools==4.41.0
fourletterphat==0.1.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpiozero==1.6.2
grpcio==1.56.2
h5py==3.9.0
html5lib==1.1
humanfriendly==10.0
idna==2.10
importlib-metadata==6.8.0
importlib-resources==6.0.0
isort==5.6.4
itsdangerous==1.1.0
jax==0.4.13
jedi==0.18.0
Jinja2==2.11.3
joblib==1.3.1
keras==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy-object-proxy==0.0.0
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.0.post2
llvmlite==0.40.1
logilab-common==1.8.1
lxml==4.6.3
madmom==0.16.1
Markdown==3.4.3
MarkupSafe==1.1.1
matplotlib==3.7.2
mccabe==0.6.1
microdotphat==0.2.1
mido==1.2.10
ml-dtypes==0.2.0
mock==5.1.0
mote==0.0.4
motephat==0.0.3
mpmath==1.3.0
msgpack==1.0.5
mypy==0.812
mypy-extensions==0.4.3
numba==0.57.1
numpy==1.23.4
oauthlib==3.1.0
onnxruntime==1.15.1
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pantilthat==0.0.7
parso==0.8.1
pexpect==4.8.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera2==0.3.12
pidng==4.0.9
piexif==1.1.3
piglow==1.2.5
pigpio==1.78
Pillow==8.1.2
pooch==1.6.0
protobuf==3.19.6
psutil==5.8.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycairo==1.16.2
pycparser==2.21
pycups==2.0.1
pygame==1.9.6
Pygments==2.7.1
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.7.2
PyOpenGL==3.1.5
pyOpenSSL==20.0.1
pyparsing==3.0.9
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyserial==3.5b0
pysmbc==1.0.23
python-apt==2.2.1
python-dateutil==2.8.2
python-prctl==1.7
PyYAML==6.0.1
rainbowhat==0.1.0
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.0.0
responses==0.12.1
roman==2.0.0
RPi.GPIO==0.7.0
rsa==4.9
RTIMULib==7.2.1
scikit-learn==1.3.0
scipy==1.11.1
scrollphat==0.0.7
scrollphathd==1.2.1
Send2Trash==1.6.0b1
sense-hat==2.4.0
simplejpeg==1.6.4
simplejson==3.17.2
six==1.16.0
skywriter==0.0.7
sn3218==1.2.7
soundfile==0.12.1
soupsieve==2.2.1
soxr==0.3.5
spidev==3.5
ssh-import-id==5.10
sympy==1.12
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow @ https://github.com/PINTO0309/Tensorflow-bin/releases/download/v2.10.0/tensorflow-2.10.0-cp39-none-linux_aarch64.whl#sha256=a98328652111bb84513d9f3a99444a1c9b778be454d3744e56f4eb65d88773a6
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
thonny==4.0.1
threadpoolctl==3.2.0
toml==0.10.1
touchphat==0.0.1
twython==3.8.2
typed-ast==1.4.2
typing_extensions==4.7.1
unicornhathd==0.0.4
urllib3==1.26.5
v4l2-python3==0.3.2
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.16.2
